LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointermove_after_pointerover_target_removed.tentative-expected.txt


FAIL "pointermove" should not be fired when the preceding "pointerover" caused removing its target assert_equals: expected (undefined) undefined but got (object) Element node <div id="parent" style="width: 32px; height: 32px;">
    ...
PASS "pointermove" should not be fired when the preceding "pointerover" caused replacing its target
PASS "pointermove" should not be fired when the preceding "pointerover" caused removing the parent of its target
PASS "pointermove" should not be fired when the preceding "pointerover" caused replacing the parent of its target
